+++ app/Http/Controllers/Admin/HallController.php
@@ -3,18 +3,20 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\Admin\UpdateHallConfigureRequest;
 use App\Services\Admin\HallService;
+use App\Services\Validation\ValidationService;
 use Illuminate\Http\Request;
-
+use Illuminate\Support\Facades\Log;
 
 class HallController extends Controller
 {
 
-    public function __construct(private HallService $hallService)
+    public function __construct(private HallService $hallService, private ValidationService $validationService)
     {}
 
     /**
-     * Создать и удалить зал.
+     * Создать зал.
      */
     public function createHall() 
     {   
@@ -22,8 +24,16 @@ class HallController extends Controller
 
         return to_route('admin_welcome');
     }
+
+    /**
+     * Удалить зал.
+     */
     public function deleteHall($id)
     {
+        $result = $this->validationService->validationId($id);
+
+        if(!$result) return;
+
         $this->hallService->deleteHall($id);
 
         return to_route('admin_welcome');
@@ -32,9 +42,12 @@ class HallController extends Controller
     /**
      * Обновление конфигурации зала.
      */
-    public function updateHallConfigure(Request $request)
+    public function updateHallConfigure(UpdateHallConfigureRequest $request)
     {
-        $data = $this->hallService->updateHallConfigure($request);
+        $result = $request->validated();
+        if(!$result) return;
+
+        $data = $this->hallService->updateHallConfigure($request->all());
         
         return response()->json(['resultUpdate' => $data]);
     }
@@ -43,6 +56,10 @@ class HallController extends Controller
      * Получение данных о местах и их количестве и рядах.
      */
     public function getDataHall($id) {
+        $result = $this->validationService->validationId($id);
+
+        if(!$result) return;
+
         $data = $this->hallService->getDataHall($id);        
 
         return response()->json(['hall' => $data['hall'], 'chairs' => $data['chairs']]);
